feat(MultiSign): allow for suites to use the same verification substeps. Identify parent suite
BREAKING CHANGE: Globally speaking with MultiSign we have redefined the structure of the verification substeps, which may occur either at the parent step level with `subSteps` (like before) or at the parent step level with the child suite(s) for verifying the various signatures. This means that implementers of a verification flow (as UI or otherwise) now need to read in the `subSteps` property or the `suites` property of the `verificationProcess`.